# Improve performance of ResourceManager::getResourceBySha1()