This adds an index to the `PersistentResource` table drastically
improving the performance of `ResourceManager::getResourceBySha1()`
calls (= public API).