@bwaidelich 2018-06-22
Some performance tests I did locally (MacBookPro with 10k Resources):

`ResourceManager::getResourceBySha1()` 1000 iterations:

Without index:

```First: 0.0597s, last: 0.0376s, average: 0.0441s, total: 44.136s```

With index:

```First: 0.0106s, last: 0.001s, average: 0.0003s, total: 0.330s```