@bwaidelich 2018-06-22
Pew.. I started with 3.3 but considered it less of a (critical) bug and more of a performance optimization, but well..
Postgres good point!