@kdambekalns 2018-06-22
Well, being 133,7454545 times slower is at least wanting to be considered a bug :)